+++ tests/CHANGELOG.md
@@ -3,6 +3,140 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.1.15](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.14...react-native-firebase-tests@7.1.15) (2020-06-10)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.14](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.13...react-native-firebase-tests@7.1.14) (2020-06-10)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.13](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.12...react-native-firebase-tests@7.1.13) (2020-06-03)
+
+
+### Bug Fixes
+
+* **firestore:**  update isEqual API to check collection path equality ([#3738](https://github.com/invertase/react-native-firebase/issues/3738)) ([405e040](https://github.com/invertase/react-native-firebase/commit/405e04009c1550dc6897b207ae3a63ad274c6de5))
+
+
+
+
+
+## [7.1.12](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.11...react-native-firebase-tests@7.1.12) (2020-06-03)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.11](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.10...react-native-firebase-tests@7.1.11) (2020-06-03)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.10](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.9...react-native-firebase-tests@7.1.10) (2020-06-03)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.9](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.8...react-native-firebase-tests@7.1.9) (2020-06-01)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.8](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.7...react-native-firebase-tests@7.1.8) (2020-05-29)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.7](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.6...react-native-firebase-tests@7.1.7) (2020-05-29)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.6](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.5...react-native-firebase-tests@7.1.6) (2020-05-29)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.5](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.4...react-native-firebase-tests@7.1.5) (2020-05-29)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.4](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.3...react-native-firebase-tests@7.1.4) (2020-05-29)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.3](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.2...react-native-firebase-tests@7.1.3) (2020-05-29)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.2](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.1...react-native-firebase-tests@7.1.2) (2020-05-28)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+## [7.1.1](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.1.0...react-native-firebase-tests@7.1.1) (2020-05-28)
+
+**Note:** Version bump only for package react-native-firebase-tests
+
+
+
+
+
+# [7.1.0](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.0.2...react-native-firebase-tests@7.1.0) (2020-05-22)
+
+
+### Features
+
+* update native Firebase SDK versions ([#3663](https://github.com/invertase/react-native-firebase/issues/3663)) ([4db9dbc](https://github.com/invertase/react-native-firebase/commit/4db9dbc3ec20bf96de0efad15000f00b41e4a799))
+
+
+
+
+
 ## [7.0.2](https://github.com/invertase/react-native-firebase/compare/react-native-firebase-tests@7.0.1...react-native-firebase-tests@7.0.2) (2020-05-15)
 
 **Note:** Version bump only for package react-native-firebase-tests